Rename playback.exoplayer to playback.media3-exoplayer
