+++ app/build.gradle.kts
@@ -92,8 +92,8 @@ val versionTxt by tasks.registering {
 dependencies {
 	// Jellyfin
 	implementation(projects.playback.core)
-	implementation(projects.playback.exoplayer)
 	implementation(projects.playback.jellyfin)
+	implementation(projects.playback.media3Exoplayer)
 	implementation(projects.playback.media3Session)
 	implementation(projects.preference)
 	implementation(libs.jellyfin.apiclient)
