+++ playback/media3-exoplayer/build.gradle.kts
@@ -37,7 +37,6 @@ dependencies {
 	implementation(libs.androidx.media3.exoplayer.hls)
 	implementation(libs.jellyfin.androidx.media3.ffmpeg.decoder)
 	implementation(libs.androidx.media3.ui)
-	implementation(libs.androidx.media3.session)
 
 	// Logging
 	implementation(libs.timber)
