+++ playback/media3-exoplayer/src/main/AndroidManifest.xml
