+++ playback/media3-exoplayer/src/main/kotlin/ExoPlayerAudioPipeline.kt
@@ -1,4 +1,4 @@
-package org.jellyfin.playback.exoplayer
+package org.jellyfin.playback.media3.exoplayer
 
 import android.media.audiofx.LoudnessEnhancer
 import timber.log.Timber
