+++ playback/media3-exoplayer/src/main/kotlin/ExoPlayerPlugin.kt
@@ -1,4 +1,4 @@
-package org.jellyfin.playback.exoplayer
+package org.jellyfin.playback.media3.exoplayer
 
 import android.content.Context
 import org.jellyfin.playback.core.plugin.playbackPlugin
